Add Price Tax Logic

+++ src/main/java/com/coway/trust/biz/sales/eSVM/impl/eSVMApiServiceImpl.java
@@ -255,8 +255,8 @@ public class eSVMApiServiceImpl
         // rtn.setSrvPromoId(smqDet.get("srvPromoId"));
         rtn.setSrvMemPacAmt( Integer.parseInt( smqDet.get( "srvMemPacAmt" ).toString() ) );
         rtn.setSrvMemBsAmt( Integer.parseInt( smqDet.get( "srvMemBsAmt" ).toString() ) );
-        rtn.setPaymentAmt( Integer.parseInt( smqDet.get( "paymentAmt" ).toString() ) );
-        rtn.setPackageTaxPrice( Integer.parseInt( smqDet.get( "srvMemPacTxs" ).toString() ) );
+        rtn.setPaymentAmt( Double.parseDouble ( smqDet.get( "paymentAmt" ).toString() ) );
+        rtn.setPackageTaxPrice( Double.parseDouble( smqDet.get( "srvMemPacTxs" ).toString() ) );
         rtn.setTaxRate( CommonUtils.nvl( eSVMApiMapper.getTaxRate() ) );
         // Get SMQ Filter details (SAL0094D)
         // EgovMap smqFilterDet = new EgovMap();
